Fix type of splitAt
--HG--
extra : convert_revision : c9ae2e42fa59ea9b9e03cb58838928b3344baea2

+++ Data/Text/Lazy.hs
@@ -577,12 +577,12 @@ dropWhile p t0 = dropWhile' t0
 -- | /O(n)/ 'splitAt' @n t@ returns a pair whose first element is a
 -- prefix of @t@ of length @n@, and whose second is the remainder of
 -- the string. It is equivalent to @('take' n t, 'drop' n t)@.
-splitAt :: Int -> Text -> (Text, Text)
+splitAt :: Int64 -> Text -> (Text, Text)
 splitAt = loop
   where loop _ Empty      = (empty, empty)
         loop n t | n <= 0 = (empty, t)
         loop n (Chunk t ts)
-             | n < len   = let (ts',ts'') = T.splitAt n t
+             | n < len   = let (ts',ts'') = T.splitAt (fromIntegral n) t
                            in (Chunk ts' Empty, Chunk ts'' Empty)
              | otherwise = let (ts',ts'') = loop (n - len) ts
                            in (Chunk t ts', ts'')
